+++ WebApiTerra1000/Controllers/ContragentControllerV2.cs
@@ -5,9 +5,9 @@ using DataCore.DAL.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.SqlClient;
-using Microsoft.Extensions.Logging;
 using NHibernate;
 using System;
+using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 using System.Xml.Linq;
@@ -21,7 +21,8 @@ namespace WebApiTerra1000.Controllers
     {
         #region Constructor and destructor
 
-        public ContragentControllerV2(ILogger<ContragentControllerV2> logger, ISessionFactory sessionFactory) : base(logger, sessionFactory)
+        //public ContragentControllerV2(ILogger<ContragentControllerV2> logger, ISessionFactory sessionFactory) : base(logger, sessionFactory)
+        public ContragentControllerV2(ISessionFactory sessionFactory) : base(sessionFactory)
         {
             //
         }
@@ -33,20 +34,25 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/v2/contragent/")]
-        public ContentResult GetContragent(int id, FormatType format = FormatType.Xml) =>
-            GetContragentWork(SqlQueriesV2.GetContragent, id, format);
+        public ContentResult GetContragentFromCodeIdProd(string code, int id, FormatType format = FormatType.Xml) =>
+            GetContragentFromCodeIdWork(code != null
+                ? SqlQueriesContragentsV2.GetContragentFromCodeProd : SqlQueriesContragentsV2.GetContragentFromIdProd,
+                code, id, format);
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/v2/contragent_preview/")]
-        public ContentResult GetContragentPreview(int id, FormatType format = FormatType.Xml) =>
-            GetContragentWork(SqlQueriesV2.GetContragentPreview, id, format);
+        public ContentResult GetContragentFromCodeIdPreview(string code, int id, FormatType format = FormatType.Xml) =>
+            GetContragentFromCodeIdWork(code != null
+                ? SqlQueriesContragentsV2.GetContragentFromCodePreview : SqlQueriesContragentsV2.GetContragentFromIdPreview,
+                code, id, format);
 
-        private ContentResult GetContragentWork(string url, int id, FormatType format = FormatType.Xml)
+        private ContentResult GetContragentFromCodeIdWork(string url, string code, int id, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url, new SqlParameter("ID", id));
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url,
+                    code != null ? TerraUtils.Sql.GetParametersV2(code) : TerraUtils.Sql.GetParametersV2(id));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Contragents} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
                 return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
@@ -56,24 +62,57 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/v2/contragents/")]
-        public ContentResult GetContragents(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
-            FormatType format = FormatType.Xml) =>
-            GetContragentsWork(SqlQueriesV2.GetContragents, startDate, endDate, offset, rowCount, format);
+        public ContentResult GetContragentsProd(DateTime? startDate, DateTime? endDate = null,
+            int? offset = null, int? rowCount = null, FormatType format = FormatType.Xml)
+        {
+            if (startDate != null && endDate != null && offset != null && rowCount != null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromDatesOffsetProd, startDate, endDate, offset, rowCount, format);
+            else if (startDate != null && endDate != null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromDatesProd, startDate, endDate, offset, rowCount, format);
+            else if (startDate != null && endDate == null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromStartDateProd, startDate, endDate, offset, rowCount, format);
+            return GetContragentsEmptyWork(SqlQueriesContragentsV2.GetContragentsEmptyProd, format);
+        }
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/v2/contragents_preview/")]
-        public ContentResult GetContragentsPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
-            FormatType format = FormatType.Xml) =>
-            GetContragentsWork(SqlQueriesV2.GetContragentsPreview, startDate, endDate, offset, rowCount, format);
+        public ContentResult GetContragentsPreview(DateTime? startDate, DateTime? endDate = null,
+            int? offset = null, int? rowCount = null, FormatType format = FormatType.Xml)
+        {
+            if (startDate != null && endDate != null && offset != null && rowCount != null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromDatesOffsetPreview, startDate, endDate, offset, rowCount, format);
+            else if (startDate != null && endDate != null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromDatesPreview, startDate, endDate, offset, rowCount, format);
+            else if (startDate != null && endDate == null)
+                return GetContragentsWork(SqlQueriesContragentsV2.GetContragentsFromStartDatePreview, startDate, endDate, offset, rowCount, format);
+            return GetContragentsEmptyWork(SqlQueriesContragentsV2.GetContragentsEmptyPreview, format);
+        }
 
-        private ContentResult GetContragentsWork(string url, DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
-            FormatType format = FormatType.Xml)
+        private ContentResult GetContragentsEmptyWork(string url, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url,
-                    TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url);
+                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
+                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
+                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
+            }), format);
+        }
+
+        private ContentResult GetContragentsWork(string url, DateTime? startDate = null, DateTime? endDate = null,
+            int? offset = null, int? rowCount = null, FormatType format = FormatType.Xml)
+        {
+            return Controller.RunTask(new Task<ContentResult>(() =>
+            {
+                List<SqlParameter> parameters = null;
+                if (startDate != null && endDate != null && offset != null && rowCount != null)
+                    parameters = TerraUtils.Sql.GetParametersV2(startDate, endDate, offset, rowCount);
+                else if (startDate != null && endDate != null)
+                    parameters = TerraUtils.Sql.GetParametersV2(startDate, endDate);
+                else if (startDate != null && endDate == null)
+                    parameters = TerraUtils.Sql.GetParametersV2(startDate);
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url, parameters);
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Contragents} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
                 return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);