+++ WebApiTerra1000/Utils/SqlQueriesV2.cs
@@ -37,74 +37,6 @@ SELECT[IIS].[fnGetDeliveryPlaces] (:StartDate, :EndDate, :Offset, :RowCount)
 SELECT[IIS].[fnGetDeliveryPlacesPreview] (:StartDate, :EndDate, :Offset, :RowCount)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
-        public static string GetContragent => @"
-SELECT [IIS].[fnGetContragentByID] (:ID)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetContragentPreview => @"
-SELECT [IIS].[fnGetContragentByIDPreview] (:ID)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetContragents => @"
-SELECT[IIS].[fnGetContragentChangesList] (:StartDate, :EndDate, :Offset, :RowCount)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetContragentsPreview => @"
-SELECT[IIS].[fnGetContragentChangesListPreview] (:StartDate, :EndDate, :Offset, :RowCount)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        #region Nomenclatures
-
-        public static string GetNomenclatureFromCodeProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (:code, DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclatureFromCodePreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (:code, DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclatureFromIdProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (DEFAULT, :id, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclatureFromIdPreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (DEFAULT, :id, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesEmptyProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesEmptyPreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t'); 
-        
-        public static string GetNomenclaturesFromStartDateProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (DEFAULT, DEFAULT, :start_date, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesFromStartDatePreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (DEFAULT, DEFAULT, :start_date, DEFAULT, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesFromDatesProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (DEFAULT, DEFAULT, :start_date, :end_date, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesFromDatesPreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (DEFAULT, DEFAULT, :start_date, :end_date, DEFAULT, DEFAULT)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesFromDatesOffsetProd => @"
-SELECT [IIS].[fnGetNomenclaturesV2] (DEFAULT, DEFAULT, :start_date, :end_date, :offset, :row_count)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        public static string GetNomenclaturesFromDatesOffsetPreview => @"
-SELECT [IIS].[fnGetNomenclaturesV2Preview] (DEFAULT, DEFAULT, :start_date, :end_date, :offset, :row_count)
-            ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
-        #endregion
-
         public static string GetShipment => @"
 SELECT [IIS].[GetRefShipmentsById] (:ID)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
@@ -125,10 +57,8 @@ SELECT [IIS].[GetRefShipmentsByDocDatePreview] (:StartDate,:EndDate,:Offset,:Row
 SELECT [IIS].[fnGetSummaryList] (:StartDate, :EndDate)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
 
-
         public static string GetSummaryPreview => @"
 SELECT [IIS].[fnGetSummaryListPreview] (:StartDate, :EndDate)
             ".TrimStart('\r', ' ', '\n', '\t').TrimEnd('\r', ' ', '\n', '\t');
-
     }
 }